Moved search close action to project board actions
